+++ docs/project/changelog.md
@@ -15,6 +15,16 @@ myst:
 
 # Change Log
 
+## Unreleased
+
+- {{ Fix }} Fixed a segfault error in the garbage collector due to stack
+  switching. The garbage collector would walk `tstate.current_frame` to find
+  objects it could mark as definitely alive. However, `tstate.current_frame`
+  includes some stack memory which can be evicted when the task is suspended. To
+  fix, we null out `tstate.current_frame` while the task is suspended and
+  restore it when we resume.
+  {pr}`6466`
+
 ## Version 314.0.7
 
 _September 14, 2026_